+++ src/main/java/com/github/jlangch/venice/impl/VeniceInterpreter.java
@@ -87,6 +87,7 @@ import com.github.jlangch.venice.impl.util.CallStack;
 import com.github.jlangch.venice.impl.util.Inspector;
 import com.github.jlangch.venice.impl.util.MeterRegistry;
 import com.github.jlangch.venice.impl.util.WithCallStack;
+import com.github.jlangch.venice.impl.util.ArityExceptions.FnType;
 import com.github.jlangch.venice.impl.util.reflect.ReflectionAccessor;
 import com.github.jlangch.venice.javainterop.AcceptAllInterceptor;
 import com.github.jlangch.venice.javainterop.IInterceptor;
@@ -351,7 +352,7 @@ public class VeniceInterpreter implements Serializable  {
 						if (numArgs != 2 || numArgs != 3) {
 							// only create callstack when needed!
 							try (WithCallStack cs = new WithCallStack(new CallFrame("if", a0.getMeta()))) {
-								ArityExceptions.assertArity("if", args, 2, 3);
+								ArityExceptions.assertArity("if", FnType.SpecialForm, args, 2, 3);
 							}
 						}
 						final VncVal cond = evaluate(args.first(), env);
@@ -366,7 +367,7 @@ public class VeniceInterpreter implements Serializable  {
 						if (args.isEmpty()) {
 							// only create callstack when needed!
 							try (WithCallStack cs = new WithCallStack(new CallFrame("let", a0.getMeta()))) {
-								ArityExceptions.assertMinArity("let", args, 1);					
+								ArityExceptions.assertMinArity("let", FnType.SpecialForm, args, 1);					
 							}
 						}
 						env = new Env(env);  // let introduces a new environment
@@ -390,7 +391,7 @@ public class VeniceInterpreter implements Serializable  {
 					if (args.size() != 1) {
 						// only create callstack when needed!
 						try (WithCallStack cs = new WithCallStack(new CallFrame("quasiquote", a0.getMeta()))) {
-							ArityExceptions.assertArity("quasiquote", args, 1);
+							ArityExceptions.assertArity("quasiquote", FnType.SpecialForm, args, 1);
 						}
 					}
 					orig_ast = quasiquote(args.first());
@@ -400,7 +401,7 @@ public class VeniceInterpreter implements Serializable  {
 					if (args.size() != 1) {
 						// only create callstack when needed!
 						try (WithCallStack cs = new WithCallStack(new CallFrame("quote", a0.getMeta()))) {
-							ArityExceptions.assertArity("quote", args, 1);
+							ArityExceptions.assertArity("quote", FnType.SpecialForm, args, 1);
 						}
 					}
 					return args.first();
@@ -409,7 +410,7 @@ public class VeniceInterpreter implements Serializable  {
 						if (args.size() < 2) {
 							// only create callstack when needed!
 							try (WithCallStack cs = new WithCallStack(new CallFrame("loop", a0.getMeta()))) {
-								ArityExceptions.assertMinArity("loop", args, 2);
+								ArityExceptions.assertMinArity("loop", FnType.SpecialForm, args, 2);
 							}
 						}
 						recursionPoint = null;
@@ -835,7 +836,7 @@ public class VeniceInterpreter implements Serializable  {
 			final AtomicInteger expandedMacrosCounter
 	) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("macroexpand", args, 1);
+			ArityExceptions.assertArity("macroexpand", FnType.SpecialForm, args, 1);
 			final VncVal ast = evaluate(args.first(), env);
 			return macroexpand(ast, env, expandedMacrosCounter);
 		}		
@@ -1003,7 +1004,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal defmacro_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertMinArity("defmacro", args, 2);
+			ArityExceptions.assertMinArity("defmacro", FnType.SpecialForm, args, 2);
 			return defmacro_(args, env);
 		}
 	}
@@ -1094,7 +1095,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal def_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("def", args, 1, 2);
+			ArityExceptions.assertArity("def", FnType.SpecialForm, args, 1, 2);
 			final VncSymbol name = validateSymbolWithCurrNS(
 										qualifySymbolWithCurrNS(
 												evaluateSymbolMetaData(args.first(), env)),
@@ -1110,7 +1111,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal defonce_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("defonce", args, 1, 2);
+			ArityExceptions.assertArity("defonce", FnType.SpecialForm, args, 1, 2);
 			final VncSymbol name = validateSymbolWithCurrNS(
 										qualifySymbolWithCurrNS(
 												evaluateSymbolMetaData(args.first(), env)),
@@ -1126,7 +1127,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal def_dynamic_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("def-dynamic", args, 1, 2);
+			ArityExceptions.assertArity("def-dynamic", FnType.SpecialForm, args, 1, 2);
 			final VncSymbol name = validateSymbolWithCurrNS(
 										qualifySymbolWithCurrNS(
 												evaluateSymbolMetaData(args.first(), env)),
@@ -1142,7 +1143,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal deftype_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("deftype", args, 2, 3);
+			ArityExceptions.assertArity("deftype", FnType.SpecialForm, args, 2, 3);
 			final VncKeyword type = Coerce.toVncKeyword(evaluate(args.first(), env));
 			final VncVector fields = Coerce.toVncVector(args.second());
 			final VncFunction validationFn = args.size() == 3
@@ -1155,7 +1156,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal deftypeQ_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("deftype?", args, 1);
+			ArityExceptions.assertArity("deftype?", FnType.SpecialForm, args, 1);
 			final VncVal type = evaluate(args.first(), env);
 			return VncBoolean.of(DefTypeForm.isCustomType(type, env));
 		}
@@ -1163,7 +1164,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal deftype_of_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertMinArity("deftype-of", args, 2);
+			ArityExceptions.assertMinArity("deftype-of", FnType.SpecialForm, args, 2);
 			final VncKeyword type = Coerce.toVncKeyword(evaluate(args.first(), env));
 			final VncKeyword baseType = Coerce.toVncKeyword(evaluate(args.second(), env));
 			final VncFunction validationFn = args.size() == 3
@@ -1181,7 +1182,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal deftype_or_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertMinArity("deftype-or", args, 2);
+			ArityExceptions.assertMinArity("deftype-or", FnType.SpecialForm, args, 2);
 			final VncKeyword type = Coerce.toVncKeyword(evaluate(args.first(), env));
 			final VncList choiceVals = args.rest();
 
@@ -1191,7 +1192,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal deftype_create_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertMinArity(".:", args, 1);
+			ArityExceptions.assertMinArity(".:", FnType.SpecialForm, args, 1);
 			final List<VncVal> evaluatedArgs = new ArrayList<>();
 			for(VncVal v : args) {
 				evaluatedArgs.add(evaluate(v, env));
@@ -1202,7 +1203,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal defmulti_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("defmulti", args, 2);
+			ArityExceptions.assertArity("defmulti", FnType.SpecialForm, args, 2);
 			final VncSymbol name =  validateSymbolWithCurrNS(
 										qualifySymbolWithCurrNS(
 												evaluateSymbolMetaData(args.first(), env)),
@@ -1230,7 +1231,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal defmethod_(final CallFrame callframe, final VncList args, final Env env, final VncVal meta) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertMinArity("defmethod", args, 2);
+			ArityExceptions.assertMinArity("defmethod", FnType.SpecialForm, args, 2);
 			final VncSymbol multiFnName = qualifySymbolWithCurrNS(
 											Coerce.toVncSymbol(args.first()));
 			final VncVal multiFnVal = env.getGlobalOrNull(multiFnName);
@@ -1266,7 +1267,7 @@ public class VeniceInterpreter implements Serializable  {
 	private VncVal ns_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
 			specialFormCallValidation("ns");
-			ArityExceptions.assertArity("ns", args, 1);
+			ArityExceptions.assertArity("ns", FnType.SpecialForm, args, 1);
 
 			final VncVal name = args.first();
 			final VncSymbol ns = Types.isVncSymbol(name)
@@ -1293,7 +1294,7 @@ public class VeniceInterpreter implements Serializable  {
 	private VncVal ns_remove_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
 			specialFormCallValidation("ns-remove");
-			ArityExceptions.assertArity("ns-remove", args, 1);
+			ArityExceptions.assertArity("ns-remove", FnType.SpecialForm, args, 1);
 
 			final VncSymbol ns = Namespaces.lookupNS(args.first(), env);
 			if (Namespaces.isSystemNS(ns.getName()) && sealedSystemNS.get()) {
@@ -1311,7 +1312,7 @@ public class VeniceInterpreter implements Serializable  {
 	private VncVal ns_unmap_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
 			specialFormCallValidation("ns-unmap");
-			ArityExceptions.assertArity("ns-unmap", args, 2);
+			ArityExceptions.assertArity("ns-unmap", FnType.SpecialForm, args, 2);
 
 			final VncSymbol ns = Namespaces.lookupNS(args.first(), env);
 			if (Namespaces.isSystemNS(ns.getName()) && sealedSystemNS.get()) {
@@ -1329,7 +1330,7 @@ public class VeniceInterpreter implements Serializable  {
 	private VncVal ns_list_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
 			specialFormCallValidation("ns-list");
-			ArityExceptions.assertArity("ns-list", args, 1);
+			ArityExceptions.assertArity("ns-list", FnType.SpecialForm, args, 1);
 
 			final VncSymbol ns = Types.isVncSymbol(args.first())
 									? (VncSymbol)args.first()
@@ -1351,7 +1352,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal import_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertMinArity("import", args, 0);
+			ArityExceptions.assertMinArity("import", FnType.SpecialForm, args, 0);
 			args.forEach(i -> Namespaces
 								.getCurrentNamespace()
 								.getJavaImports()
@@ -1381,7 +1382,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal namespace_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("namespace", args, 1);
+			ArityExceptions.assertArity("namespace", FnType.SpecialForm, args, 1);
 			final VncVal val = evaluate(args.first(), env);
 			if (val instanceof INamespaceAware) {
 				return new VncString(((INamespaceAware)val).getNamespace());
@@ -1396,7 +1397,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal resolve_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("resolve", args, 1);
+			ArityExceptions.assertArity("resolve", FnType.SpecialForm, args, 1);
 			specialFormCallValidation("resolve");
 			return env.getOrNil(Coerce.toVncSymbol(evaluate(args.first(), env)));
 		}
@@ -1404,7 +1405,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal var_get_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("var-get", args, 1);
+			ArityExceptions.assertArity("var-get", FnType.SpecialForm, args, 1);
 			specialFormCallValidation("var-get");
 			final VncSymbol sym = Types.isVncSymbol(args.first())
 									? (VncSymbol)args.first()
@@ -1415,7 +1416,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal var_ns_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("var-ns", args, 1);
+			ArityExceptions.assertArity("var-ns", FnType.SpecialForm, args, 1);
 			specialFormCallValidation("var-ns");
 			final VncSymbol sym = Types.isVncSymbol(args.first())
 									? (VncSymbol)args.first()
@@ -1427,7 +1428,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal var_name_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("var-name", args, 1);
+			ArityExceptions.assertArity("var-name", FnType.SpecialForm, args, 1);
 			specialFormCallValidation("var-name");
 			final VncSymbol sym = Types.isVncSymbol(args.first())
 									? (VncSymbol)args.first()
@@ -1438,7 +1439,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal var_localQ_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("var-local?", args, 1);
+			ArityExceptions.assertArity("var-local?", FnType.SpecialForm, args, 1);
 			final VncSymbol sym = Types.isVncSymbol(args.first())
 									? (VncSymbol)args.first()
 									: Coerce.toVncSymbol(evaluate(args.first(), env));
@@ -1448,7 +1449,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal var_thread_localQ_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("var-thread-local?", args, 1);
+			ArityExceptions.assertArity("var-thread-local?", FnType.SpecialForm, args, 1);
 			final VncSymbol sym = Types.isVncSymbol(args.first())
 									? (VncSymbol)args.first()
 									: Coerce.toVncSymbol(evaluate(args.first(), env));
@@ -1458,7 +1459,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal var_globalQ_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("var-global?", args, 1);
+			ArityExceptions.assertArity("var-global?", FnType.SpecialForm, args, 1);
 			final VncSymbol sym = Types.isVncSymbol(args.first())
 									? (VncSymbol)args.first()
 									: Coerce.toVncSymbol(evaluate(args.first(), env));
@@ -1468,7 +1469,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal setBANG_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("set!", args, 2);
+			ArityExceptions.assertArity("set!", FnType.SpecialForm, args, 2);
 			specialFormCallValidation("set!");
 	
 			final VncSymbol sym = Types.isVncSymbol(args.first())
@@ -1498,7 +1499,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal inspect_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("inspect", args, 1);
+			ArityExceptions.assertArity("inspect", FnType.SpecialForm, args, 1);
 			specialFormCallValidation("inspect");
 			final VncSymbol sym = Coerce.toVncSymbol(evaluate(args.first(), env));
 			return Inspector.inspect(env.get(sym));
@@ -1519,7 +1520,7 @@ public class VeniceInterpreter implements Serializable  {
 	
 	private VncVal doc_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("doc", args, 1);
+			ArityExceptions.assertArity("doc", FnType.SpecialForm, args, 1);
 			final VncString doc = DocForm.doc(args.first(), env);
 			evaluate(VncList.of(new VncSymbol("println"), doc), env);
 			return Nil;
@@ -1528,7 +1529,7 @@ public class VeniceInterpreter implements Serializable  {
 	
 	private VncVal print_highlight_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("print-highlight", args, 1);
+			ArityExceptions.assertArity("print-highlight", FnType.SpecialForm, args, 1);
 			final VncString form = DocForm.highlight(Coerce.toVncString(args.first()), env);
 			evaluate(VncList.of(new VncSymbol("println"), form), env);
 			return Nil;
@@ -1550,7 +1551,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal eval_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertMinArity("eval", args, 0);
+			ArityExceptions.assertMinArity("eval", FnType.SpecialForm, args, 0);
 			specialFormCallValidation("eval");
 			final Namespace ns = Namespaces.getCurrentNamespace();
 			try {
@@ -1564,7 +1565,7 @@ public class VeniceInterpreter implements Serializable  {
 	
 	private VncVal dorun_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("dorun", args, 2);
+			ArityExceptions.assertArity("dorun", FnType.SpecialForm, args, 2);
 			
 			final long count = Coerce.toVncLong(args.first()).getValue();
 			if (count <= 0) return Nil;
@@ -1593,7 +1594,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal dobench_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("dobench", args, 2);
+			ArityExceptions.assertArity("dobench", FnType.SpecialForm, args, 2);
 			
 			try {
 				final long count = Coerce.toVncLong(args.first()).getValue();
@@ -1624,7 +1625,7 @@ public class VeniceInterpreter implements Serializable  {
 
 	private VncVal locking_(final CallFrame callframe, final VncList args, final Env env) {
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertMinArity("locking", args, 2);
+			ArityExceptions.assertMinArity("locking", FnType.SpecialForm, args, 2);
 			
 			final VncVal mutex = evaluate(args.first(), env);
 	
@@ -1659,7 +1660,7 @@ public class VeniceInterpreter implements Serializable  {
 		// multi arity:   (fn name? ([params*] condition-map? expr*)+ )
 
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertMinArity("fn", args, 1);
+			ArityExceptions.assertMinArity("fn", FnType.SpecialForm, args, 1);
 	
 			VncSymbol name;
 			int argPos;
@@ -1747,7 +1748,7 @@ public class VeniceInterpreter implements Serializable  {
 		// See:  - https://smartbear.com/learn/code-profiling/fundamentals-of-performance-profiling/
 		//       - https://support.smartbear.com/aqtime/docs/profiling-with/profile-various-apps/recursive-routines.html
 		try (WithCallStack cs = new WithCallStack(callframe)) {
-			ArityExceptions.assertArity("prof", args, 1, 2, 3);
+			ArityExceptions.assertArity("prof", FnType.SpecialForm, args, 1, 2, 3);
 
 			if (Types.isVncKeyword(args.first())) {
 				final VncKeyword cmd = (VncKeyword)args.first();
@@ -2029,7 +2030,8 @@ public class VeniceInterpreter implements Serializable  {
 					if (args.size() < getFixedArgsCount()) {
 							throw new ArityException(
 									ArityExceptions.formatVariadicArityExMsg(
-										getQualifiedName(),
+										getQualifiedName(), 
+										macro ? FnType.Macro : FnType.Function,
 										args.size(), 
 										getFixedArgsCount(),
 										getArgLists()));
@@ -2038,7 +2040,8 @@ public class VeniceInterpreter implements Serializable  {
 				else if (args.size() != getFixedArgsCount()) {
 						throw new ArityException(
 								ArityExceptions.formatArityExMsg(
-									getQualifiedName(),
+									getQualifiedName(), 
+									macro ? FnType.Macro : FnType.Function,
 									args.size(), 
 									getFixedArgsCount(),
 									getArgLists()));