+++ ChangeLog.md
@@ -24,6 +24,7 @@ and this project adheres to
 
 - word wrap with function names in PDF cheatsheet
 - `shell/process-info` to handle Java Optional type better
+- callstack for macro arity exception